refactor: build_json_bed_files.py renamed to catalog_utils.py and moved to utils package. Use of default catalog files placed into data folder (#40)
