--- categories/PR/PR_risk_genes_GRCh38.bed
@@ -1,78 +1,0 @@
-chr1	17018664	17054151	SDHB
-chr1	45329163	45340893	MUTYH
-chr1	55039445	55064852	PCSK9
-chr1	68428822	68451103	RPE65
-chr1	115698038	115837921	CASQ2
-chr1	156082572	156140081	LMNA
-chr1	161314381	161363206	SDHC
-chr1	201039512	201112451	CACNA1S
-chr1	201359008	201377764	TNNT2
-chr1	237042184	237833988	RYR2
-chr2	21001429	21044073	APOB
-chr2	47403067	47663146	MSH2
-chr2	47695530	47810063	MSH6
-chr2	96248500	96266047	TMEM127
-chr2	178525989	178830802	TTN
-chr2	188974373	189012746	COL3A1
-chr2	219418377	219426735	DES
-chr3	10141778	10153676	VHL
-chr3	14125015	14143680	TMEM43
-chr3	15601341	15722311	BTD
-chr3	30606478	30694249	TGFBR2
-chr3	36993226	37050896	MLH1
-chr3	38548057	38649743	SCN5A
-chr3	46835110	46882178	MYL3
-chr3	52451100	52454041	TNNC1
-chr5	112707518	112846239	APC
-chr6	7541326	7586717	DSP
-chr6	26087226	26098343	HFE
-chr6	123216339	123637189	TRDN
-chr7	5970925	6009130	PMS2
-chr7	128830406	128859274	FLNC
-chr7	150944961	150978321	KCNH2
-chr7	151556124	151877214	PRKAG2
-chr9	99104038	99154192	TGFBR1
-chr9	127811130	127854773	ENG
-chr9	132891348	132946874	TSC1
-chr10	43077064	43130351	RET
-chr10	86755786	86932825	BMPR1A
-chr10	87862638	87971930	PTEN
-chr10	88934822	88991339	ACTA2
-chr10	110644336	110839471	RBM20
-chr10	119651380	119677819	BAG3
-chr11	2444654	2849105	KCNQ1
-chr11	32387775	32435564	WT1
-chr11	47331406	47352702	MYBPC3
-chr11	61430042	61446839	SDHAF2
-chr11	64803510	64811294	MEN1
-chr11	112086824	112120016	SDHD
-chr12	32790745	32896798	PKP2
-chr12	51906908	51923361	ACVRL1
-chr12	110909996	111052434	MYL2
-chr12	120978543	121002512	HNF1A
-chr13	32315086	32400268	BRCA2
-chr13	48303744	48599436	RB1
-chr13	51930436	52012125	ATP7B
-chr14	23412740	23435660	MYH7
-chr14	65006174	65102695	MAX
-chr15	34790107	34795589	ACTC1
-chr15	48408313	48645721	FBN1
-chr15	63042620	63071915	TPM1
-chr15	67063763	67195173	SMAD3
-chr16	2047967	2089491	TSC2
-chr16	15703135	15858438	MYH11
-chr16	23603160	23641321	PALB2
-chr17	7661779	7687546	TP53
-chr17	43044295	43170245	BRCA1
-chr17	80101533	80119881	GAA
-chr18	31058545	31102600	DSC2
-chr18	31497624	31551351	DSG2
-chr18	31557009	31598833	TTR
-chr18	51028528	51085045	SMAD4
-chr19	1177558	1228431	STK11
-chr19	11089418	11133820	LDLR
-chr19	38433691	38595273	RYR1
-chr19	55151767	55157773	TNNI3
-chr22	29603553	29698598	NF2
-chrX	38327598	38422908	OTC
-chrX	101393273	101408012	GLA
