--- categories/RR/RR_risk_genes_GRCh38.bed
@@ -1,113 +1,0 @@
-chr1	21509397	21578410	ALPL
-chr1	26432282	26471306	DHDDS
-chr1	45500300	45513382	MMACHC
-chr1	53196792	53214197	CPT2
-chr1	75724431	75787575	ACADM
-chr1	155234452	155244699	GBA1
-chr1	171090901	171117819	FMO3
-chr1	215622891	216423448	USH2A
-chr2	151485336	151734487	NEB
-chr2	169127109	169362534	LRP2
-chr2	218781749	218815293	CYP27A1
-chr2	227683763	227718028	SLC19A3
-chr2	240868824	240880502	AGXT
-chr3	14145145	14178621	XPC
-chr3	15601341	15722311	BTD
-chr3	43354859	43691594	ANO10
-chr3	48564073	48595329	COL7A1
-chr3	81489703	81761645	GBE1
-chr3	133746040	133796641	TF
-chr3	149129638	149173732	HPS3
-chr3	150926163	150972727	CLRN1
-chr4	986997	1004564	IDUA
-chr4	5542772	5709548	EVC2
-chr4	15469865	15601552	CC2D2A
-chr4	177430774	177442437	AGA
-chr5	70925030	70953942	SMN1
-chr5	71579531	71658706	MCCC2
-chr5	149960758	149993455	SLC26A2
-chr6	32038327	32041644	CYP21A2
-chr6	32041153	32115334	TNXB
-chr6	49430360	49463253	MMUT
-chr6	51615299	52087613	PKHD1
-chr6	80106647	80346270	BCKDHB
-chr6	87513459	87590028	RARS2
-chr6	135283407	135498434	AHI1
-chr7	66075800	66094697	ASL
-chr7	107660828	107717809	SLC26A4
-chr7	107891162	107931730	DLD
-chr7	117287120	117715971	CFTR
-chr7	143316111	143352083	CLCN1
-chr8	6406592	6648508	MCPH1
-chr8	86553977	86743675	CNGB3
-chr9	34638133	34651035	GALT
-chr9	69035751	69079076	FXN
-chr9	95099054	95426796	FANCC
-chr9	101420560	101449664	ALDOB
-chr9	105558122	105653820	FKTN
-chr9	108866898	108934328	ELP1
-chr10	53802771	55627942	PCDH15
-chr10	70597348	70602759	PRF1
-chr10	98410939	98446963	HPS1
-chr11	5225464	5229395	HBB
-chr11	6390440	6394998	SMPD1
-chr11	17392498	17476894	ABCC8
-chr11	61392393	61398866	TMEM216
-chr11	66510606	66533613	BBS1
-chr11	71428193	71452868	DHCR7
-chr11	89177875	89295759	TYR
-chr11	103109410	103479863	DYNC2H1
-chr11	108116695	108148957	ACAT1
-chr11	119023751	119030906	SLC37A4
-chr12	57762334	57768986	CYP27B1
-chr12	66347431	67069162	GRIP1
-chr12	88049016	88142099	CEP290
-chr12	101745499	101830959	GNPTAB
-chr12	102836889	102958410	PAH
-chr12	109573255	109598125	MVK
-chr13	20187463	20192938	GJB2
-chr13	50909747	51024120	RNASEH2B
-chr13	51930436	52012125	ATP7B
-chr14	91271323	91417844	CCDC88C
-chr15	27754875	28099315	OCA2
-chr15	72340924	72376420	HEXA
-chr15	74337759	74367646	CYP11A1
-chr15	80152490	80186946	FAH
-chr15	89305198	89334861	POLG
-chr15	90717346	90816166	BLM
-chr16	172876	173710	HBA2
-chr16	176680	177522	HBA1
-chr16	2275881	2340746	ABCA3
-chr16	8788823	8862534	PMM2
-chr16	56465640	56582667	BBS2
-chr17	3472374	3503405	ASPA
-chr17	4897771	4934438	CHRNE
-chr17	7217125	7225266	ACADVL
-chr17	42900797	42914438	G6PC1
-chr17	80101533	80119881	GAA
-chr19	7522624	7534009	MCOLN1
-chr19	35825372	35869287	NPHS1
-chr19	45349837	45370918	ERCC2
-chr19	46746046	46776988	FKRP
-chr21	43053191	43076943	CBS
-chr21	44285838	44298648	AIRE
-chr22	42058334	42070842	NAGA
-chr22	50059391	50085426	MLC1
-chr22	50523568	50526461	SCO2
-chr22	50622754	50628173	ARSA
-chrX	10445310	10833654	MID1
-chrX	18639688	18672108	RS1
-chrX	25003694	25016420	ARX
-chrX	30304206	30309390	NR0B1
-chrX	31097677	33339609	DMD
-chrX	38269163	38327544	RPGR
-chrX	38327598	38422908	OTC
-chrX	101393273	101408012	GLA
-chrX	103773718	103792619	PLP1
-chrX	139530739	139563459	F9
-chrX	147911919	147951125	FMR1
-chrX	148500617	149000663	AFF2
-chrX	153687926	153696588	SLC6A8
-chrX	153724856	153744755	ABCD1
-chrX	153861514	153886173	L1CAM
-chrX	154835788	155026940	F8
