+++ examples/config_example.json
@@ -16,14 +16,14 @@
       "GRCh37": "/mnt/lustre/scratch/CBRA/data/indexed_genomes/bwa/hs37d5/hs37d5.fa",
       "GRCh38": "/mnt/lustre/scratch/CBRA/data/indexed_genomes/bwa/hg38_no_alt/hg38_no_alt.fa"
     },
-    "gene_to_phenotype_file": "/home/jpflorido/GBPA/projects/SFtool/categories/genes_to_phenotype_20250506.txt",
+    "gene_to_phenotype_file": "/home/jpflorido/GBPA/projects/SFtool/sftool/data/categories/genes_to_phenotype_20260623.txt",
     "pharmCAT_positions_vcf": "/mnt/lustre/scratch/CBRA/projects/ACCI/secondary_findings/resources/pharmcat_positions_3.2.0.vcf"
   },
 
   "catalogs": {
-    "personal_risk_geneset": "/home/jpflorido/GBPA/projects/SFtool/categories/PR/PR_risk_genes_ACMG_SF_v3.1.csv",
-    "reproductive_risk_geneset": "/home/jpflorido/GBPA/projects/SFtool/categories/RR/RR_risk_genes_ACMG_CS_v2021.csv",
-    "reproductive_risk_geneset_STR": "/home/jpflorido/GBPA/projects/SFtool/categories/RR/RR_risk_genes_STR_ACMG_CS_v2021.csv"
+    "personal_risk_geneset": "/home/jpflorido/GBPA/projects/SFtool/sftool/data/categories/PR/PR_risk_genes_ACMG_SF_v3.1.csv",
+    "reproductive_risk_geneset": "/home/jpflorido/GBPA/projects/SFtool/sftool/data/categories/RR/RR_risk_genes_ACMG_CS_v2021.csv",
+    "reproductive_risk_geneset_STR": "/home/jpflorido/GBPA/projects/SFtool/sftool/data/categories/RR/RR_risk_genes_STR_ACMG_CS_v2021.csv"
   },
 
   "clinvar": {
