+++ sftool/data/categories/PR/PR_risk_genes_GRCh37.bed
