+++ sftool/data/categories/PR/PR_risk_genes_GRCh38.bed
@@ -1,0 +1,78 @@
+chr1	17018664	17054210	SDHB
+chr1	45329163	45340893	MUTYH
+chr1	55039445	55064852	PCSK9
+chr1	68428822	68451103	RPE65
+chr1	115697988	115837921	CASQ2
+chr1	156082572	156140085	LMNA
+chr1	161314352	161363774	SDHC
+chr1	201039512	201112464	CACNA1S
+chr1	201356606	201379773	TNNT2
+chr1	237042184	237833988	RYR2
+chr2	21001305	21044075	APOB
+chr2	47403067	47663146	MSH2
+chr2	47695530	47810109	MSH6
+chr2	96248500	96266082	TMEM127
+chr2	178525989	178830802	TTN
+chr2	188974371	189012746	COL3A1
+chr2	219417474	219426740	DES
+chr3	10141764	10153676	VHL
+chr3	14124998	14143681	TMEM43
+chr3	15601341	15722522	BTD
+chr3	30606478	30694249	TGFBR2
+chr3	36992834	37051505	MLH1
+chr3	38548057	38649743	SCN5A
+chr3	46835109	46885911	MYL3
+chr3	52451082	52454071	TNNC1
+chr5	112707452	112846239	APC
+chr6	7541320	7586717	DSP
+chr6	26087226	26098343	HFE
+chr6	123215088	123637316	TRDN
+chr7	5970925	6009130	PMS2
+chr7	128830292	128859275	FLNC
+chr7	150944064	150978394	KCNH2
+chr7	151556010	151877563	PRKAG2
+chr9	99104038	99154395	TGFBR1
+chr9	127811130	127855232	ENG
+chr9	132891348	132946874	TSC1
+chr10	43077026	43130351	RET
+chr10	86755751	86932825	BMPR1A
+chr10	87862638	87971930	PTEN
+chr10	88934815	88991339	ACTA2
+chr10	110643874	110839471	RBM20
+chr10	119650879	119677913	BAG3
+chr11	2444654	2849105	KCNQ1
+chr11	32387775	32435885	WT1
+chr11	47331406	47352702	MYBPC3
+chr11	61430032	61473614	SDHAF2
+chr11	64803500	64811294	MEN1
+chr11	112086361	112120016	SDHD
+chr12	32790745	32896819	PKP2
+chr12	51906908	51923361	ACVRL1
+chr12	110909996	110921443	MYL2
+chr12	120978543	121002512	HNF1A
+chr13	32315086	32400268	BRCA2
+chr13	48303710	48599436	RB1
+chr13	51930436	52012198	ATP7B
+chr14	23412732	23436137	MYH7
+chr14	65005398	65102695	MAX
+chr15	34790107	34795912	ACTC1
+chr15	48408016	48645791	FBN1
+chr15	63042130	63071915	TPM1
+chr15	67063763	67195173	SMAD3
+chr16	2047345	2089491	TSC2
+chr16	15703135	15858438	MYH11
+chr16	23603158	23641342	PALB2
+chr17	7661779	7687546	TP53
+chr17	43044292	43170245	BRCA1
+chr17	80101526	80119881	GAA
+chr18	31058545	31102600	DSC2
+chr18	31497624	31551351	DSG2
+chr18	31557008	31606585	TTR
+chr18	51028528	51085045	SMAD4
+chr19	1177558	1228431	STK11
+chr19	11089418	11133830	LDLR
+chr19	38433691	38595273	RYR1
+chr19	55151763	55157774	TNNI3
+chr22	29603520	29698598	NF2
+chrX	38327598	38424632	OTC
+chrX	101393273	101408012	GLA
