+++ sftool/data/categories/RR/RR_risk_genes_GRCh37.bed
