+++ sftool/data/categories/RR/RR_risk_genes_GRCh38.bed
@@ -1,0 +1,113 @@
+chr1	21508813	21579441	ALPL
+chr1	26432273	26471322	DHDDS
+chr1	45500186	45513382	MMACHC
+chr1	53196770	53214197	CPT2
+chr1	75724431	75787575	ACADM
+chr1	155231270	155245178	GBA1
+chr1	171085113	171118675	FMO3
+chr1	215621576	216423448	USH2A
+chr2	151485336	151734515	NEB
+chr2	169127109	169362780	LRP2
+chr2	218781688	218825335	CYP27A1
+chr2	227683763	227718088	SLC19A3
+chr2	240868543	240881531	AGXT
+chr3	14145145	14178767	XPC
+chr3	15601341	15722522	BTD
+chr3	43354859	43691594	ANO10
+chr3	48564073	48595596	COL7A1
+chr3	81489161	81761840	GBE1
+chr3	133745862	133796641	TF
+chr3	149129602	149173732	HPS3
+chr3	150926163	150972727	CLRN1
+chr4	986997	1008344	IDUA
+chr4	5542772	5709548	EVC2
+chr4	15469865	15602097	CC2D2A
+chr4	177430127	177442450	AGA
+chr5	70925030	70953942	SMN1
+chr5	71579531	71658706	MCCC2
+chr5	149960748	149993455	SLC26A2
+chr6	32038297	32042321	CYP21A2
+chr6	32041153	32115334	TNXB
+chr6	49430360	49463321	MMUT
+chr6	51615299	52087618	PKHD1
+chr6	80106626	80384773	BCKDHB
+chr6	87513459	87590028	RARS2
+chr6	135283407	135498434	AHI1
+chr7	66075800	66094697	ASL
+chr7	107660594	107717809	SLC26A4
+chr7	107891136	107931730	DLD
+chr7	117287120	117715971	CFTR
+chr7	143316111	143353744	CLCN1
+chr8	6406592	6648508	MCPH1
+chr8	86553977	86743675	CNGB3
+chr9	34637367	34651754	GALT
+chr9	69035730	69079076	FXN
+chr9	95099054	95426796	FANCC
+chr9	101420560	101449664	ALDOB
+chr9	105558122	105653820	FKTN
+chr9	108866898	108934414	ELP1
+chr10	53802771	55627942	PCDH15
+chr10	70596401	70602788	PRF1
+chr10	98410939	98446966	HPS1
+chr11	5225464	5229395	HBB
+chr11	6390381	6394998	SMPD1
+chr11	17392498	17476894	ABCC8
+chr11	61392393	61400408	TMEM216
+chr11	66510606	66534660	BBS1
+chr11	71428193	71452868	DHCR7
+chr11	89177875	89295759	TYR
+chr11	103109351	103479863	DYNC2H1
+chr11	108116695	108148957	ACAT1
+chr11	119023751	119030906	SLC37A4
+chr12	57762334	57768986	CYP27B1
+chr12	66347431	67069162	GRIP1
+chr12	88049016	88142099	CEP290
+chr12	101744032	101830984	GNPTAB
+chr12	102836889	102958410	PAH
+chr12	109573255	109600609	MVK
+chr13	20187463	20192980	GJB2
+chr13	50909746	51024120	RNASEH2B
+chr13	51930436	52012198	ATP7B
+chr14	91271323	91417890	CCDC88C
+chr15	27754859	28099664	OCA2
+chr15	72340919	72376420	HEXA
+chr15	74337746	74367884	CYP11A1
+chr15	80152292	80186990	FAH
+chr15	89305198	89334972	POLG
+chr15	90717340	90816166	BLM
+chr16	172021	173910	HBA2
+chr16	176660	177527	HBA1
+chr16	2275878	2340788	ABCA3
+chr16	8788823	8862534	PMM2
+chr16	56465640	56582667	BBS2
+chr17	3472374	3503405	ASPA
+chr17	4897769	4934438	CHRNE
+chr17	7217125	7225273	ACADVL
+chr17	42900722	42923502	G6PC1
+chr17	80101526	80119881	GAA
+chr19	7522611	7534009	MCOLN1
+chr19	35825372	35869287	NPHS1
+chr19	45349773	45370918	ERCC2
+chr19	46746046	46776988	FKRP
+chr21	43050900	43076943	CBS
+chr21	44285838	44298648	AIRE
+chr22	42032595	42071198	NAGA
+chr22	50059391	50085969	MLC1
+chr22	50523564	50526461	SCO2
+chr22	50622754	50628177	ARSA
+chrX	10445310	10833730	MID1
+chrX	18639688	18672108	RS1
+chrX	25003694	25016420	ARX
+chrX	30304206	30309394	NR0B1
+chrX	31097677	33339609	DMD
+chrX	38269115	38327573	RPGR
+chrX	38327598	38424632	OTC
+chrX	101393273	101408012	GLA
+chrX	103742560	103792678	PLP1
+chrX	139530731	139568549	F9
+chrX	147911858	147951125	FMR1
+chrX	148500611	149000663	AFF2
+chrX	153687914	153696591	SLC6A8
+chrX	153724495	153744755	ABCD1
+chrX	153861511	153886204	L1CAM
+chrX	154835788	155026940	F8
