+++ sftool/data/categories/RR/RR_risk_genes_STR_ACMG_CS_v2021.csv
